+++ apps/python-docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # python-docs
 
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @steamship/react@0.8.0
+
 ## 0.0.12
 
 ### Patch Changes